Stopping point pre-valueifying strings and arrays
I originally chose to represent all of my structs as pointers in LLVM because I was
unaware of the value-based instructions insertvalue and extractvalue; I thought
you had to have a pointer to a struct in order to even use it, so I just decided
to represent them that way. This makes it impossible to have an array of structs;
I can only do an array of pointers to structs, which is not what I want.
I want to treat Arrays as values where the struct of len, cap, ptr is on the stack.

I want to be able to return these from functions and have LLVM just figure out that I'm
returning a struct and use sret w/ an out arg or whatever it wants to do.

This will involve a big refactor but it will solve most prior and hopeful many future
codegen bugs