TypeSchema for Enums; stopping point due to LLVM dead-end
We have to switch to representing Enums as byte arrays in LLVM, rather
than structs.

This is because the type, for static enums, needs to change based on the
variant.

https://github.com/llvm/llvm-project/issues/107239
This LLVM issue is relevant, and though not the same thing, since our
unions are always tagged, makes a compelling case for just using the
bytes. We already memcpy them everywhere anyway